Remove RXPK & LoRa Physical (and add simulator) (#1153)
* Start simulator migration

* Remove RXPK

* Fix glitch

* Add minor fixes

* Adapt to changes

* Fix codeql

* React on comments

* Fix test

* Continue the progress

* remove uneeded imports

* Fix tests

+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/BasicsStation/JsonHandlers/LnsData.cs
@@ -70,7 +70,7 @@ namespace LoRaWan.NetworkServer.BasicsStation.JsonHandlers
 
         private static readonly IJsonProperty<Mic> MicProperty =
             JsonReader.Property("MIC", from i in JsonReader.Int32()
-                                       select new Mic(unchecked((uint)i)));
+                                       select new Mic(unchecked(i)));
 
         /*
             {
